refactor(common): move package
